wip: tests changes
