+++ tests/package-deb-gen/.gitignore
@@ -1,0 +1,1 @@
+cargo-deb.toml
